+++ Solution/Game/PhysicsComponent.h
@@ -6,13 +6,13 @@ class PhysicsComponent : public Component
 public:
 	PhysicsComponent(Entity& aEntity);
 
-	void Init(const CU::Matrix44<float>& anOrientation, const CU::Vector3<float>& aVelocity = { 0.f, 0.f, 0.f }, unsigned short aWeight = 1);
-	void Init(unsigned short aWeight = 1, const CU::Vector3<float>& aVelocity = { 0.f, 0.f, 0.f });
+	void Init(const CU::Matrix44<float>& anOrientation, const CU::Vector3<float>& aVelocity = { 0.f, 0.f, 0.f }, int aWeight = 1);
+	void Init(int aWeight = 1, const CU::Vector3<float>& aVelocity = { 0.f, 0.f, 0.f });
 	void Update(float aDeltaTime);
 
 	static eComponentType GetType();
 
-	unsigned short GetWeight() const;
+	int GetWeight() const;
 
 	void SetVelocity(const CU::Vector3<float>& aVelocity);
 	void AddVelocity(const CU::Vector3<float>& aVelocity);
@@ -24,7 +24,7 @@ public:
 
 private:
 	CU::Vector3<float> myVelocity;
-	unsigned short myWeight;
+	int myWeight;
 };
 
 inline eComponentType PhysicsComponent::GetType()
@@ -32,7 +32,7 @@ inline eComponentType PhysicsComponent::GetType()
 	return eComponentType::PHYSICS;
 }
 
-inline unsigned short PhysicsComponent::GetWeight() const
+inline int PhysicsComponent::GetWeight() const
 {
 	return myWeight;
 }
